requirements.txt
asgiref==3.8.1
Django==5.1.1
django-debug-toolbar==4.4.6
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==30.3.0
faker_food==0.3.0
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
